producer/requirements.txt
mastodon.py==1.5.0
kafka-python==2.0.2
python-dotenv==0.21.0
omegaconf==2.1.1
requests==2.28.1